Avoiding some issues around init and destroy
The proxy won't work with multi-arity overloads when one invokes
proxy-super, because it blows away all the overloads in order to refer
to the one.

Plus init is flaky anyway, since it only seems to ever be called
lazily, prior to the first request, even when setLoadOnStartup is
called with >=0. And destroy isn't called unless init is.

The lein-ring project uses ServletContextListener's for init/destroy,
but I'm not sure that will work when multiple servlets are created in
the same context.

The bottom-line is I like the predictability of init/destroy the way
we had it before, so I'm restoring that logic.